+++ src/compile-middleware.ts
@@ -19,6 +19,7 @@ import {Request, RequestHandler, Response} from 'express';
 import * as LRU from 'lru-cache';
 import * as parse5 from 'parse5';
 import {UAParser} from 'ua-parser-js';
+import * as url from 'url';
 
 import {transformResponse} from './transform-middleware';
 
@@ -62,11 +63,6 @@ function getContentType(response: Response) {
   return contentTypeHeader && parseContentType(contentTypeHeader).type;
 }
 
-function isSuccessful(response: Response) {
-  const statusCode = response.statusCode;
-  return (statusCode >= 200 && statusCode < 300);
-}
-
 // NOTE: To change the max length of the cache at runtime, just use bracket
 // notation, i.e. `babelCompileCache['max'] = 64 * 1024` for 64KB limit.
 export const babelCompileCache = LRU<string>(<LRU.Options<string>>{
@@ -75,35 +71,33 @@ export const babelCompileCache = LRU<string>(<LRU.Options<string>>{
 
 
 export function babelCompile(forceCompile: boolean): RequestHandler {
-  if (forceCompile == null) {
-    forceCompile = false;
-  }
-
   return transformResponse({
-    shouldTransform(_request: Request, response: Response) {
-      return isSuccessful(response) &&
-          compileMimeTypes.indexOf(getContentType(response)) >= 0;
+    shouldTransform(request: Request, response: Response) {
+      // We must never compile the Custom Elements ES5 Adapter or other
+      // polyfills/shims.
+      const isPolyfill = url.parse(request.url)
+                             .pathname.split('/')
+                             .includes('webcomponentsjs');
+      return !isPolyfill &&
+          compileMimeTypes.includes(getContentType(response)) &&
+          (forceCompile ||
+           browserNeedsCompilation(request.headers['user-agent']));
     },
 
-    transform(request: Request, response: Response, body: string): string /**/ {
+    transform(request: Request, response: Response, body: string): string {
       const contentType = getContentType(response);
-      const uaParser = new UAParser(request.headers['user-agent']);
-      const compile = forceCompile || needCompilation(uaParser);
-
-      if (compile) {
-        const source = body;
-        const cached = babelCompileCache.get(source);
-        if (cached !== undefined) {
-          return cached;
-        }
-        if (contentType === htmlMimeType) {
-          body = compileHtml(source, request.path);
-        }
-        if (javaScriptMimeTypes.indexOf(contentType) !== -1) {
-          body = compileScript(source);
-        }
-        babelCompileCache.set(source, body);
+      const source = body;
+      const cached = babelCompileCache.get(source);
+      if (cached !== undefined) {
+        return cached;
       }
+      if (contentType === htmlMimeType) {
+        body = compileHtml(source, request.path);
+      }
+      if (javaScriptMimeTypes.includes(contentType)) {
+        body = compileScript(source);
+      }
+      babelCompileCache.set(source, body);
 
       return body;
     },
@@ -141,14 +135,25 @@ const isInlineJavaScript = dom5.predicates.AND(
     dom5.predicates.hasTagName('script'),
     dom5.predicates.NOT(dom5.predicates.hasAttr('src')));
 
-function needCompilation(uaParser: UAParser): boolean {
-  const browser = uaParser.getBrowser();
-  const versionSplit = browser.version && browser.version.split('.');
-  const majorVersion = versionSplit ? parseInt(versionSplit[0], 10) : -1;
+export function browserNeedsCompilation(userAgent: string): boolean {
+  const browser = new UAParser(userAgent).getBrowser();
+  const versionSplit = (browser.version || '').split('.');
+  const [majorVersion, minorVersion] =
+      versionSplit.map((v) => v ? parseInt(v, 10) : -1);
 
   const supportsES2015 = (browser.name === 'Chrome' && majorVersion >= 49) ||
+      (browser.name === 'Chromium' && majorVersion >= 49) ||
+      (browser.name === 'OPR' && majorVersion >= 36) ||
+      (browser.name === 'Mobile Safari' && majorVersion >= 10) ||
       (browser.name === 'Safari' && majorVersion >= 10) ||
-      (browser.name === 'Edge' && majorVersion >= 14) ||
+      // Note: The Edge user agent uses the EdgeHTML version, not the main
+      // release version (e.g. EdgeHTML 15 corresponds to Edge 40). See
+      // https://en.wikipedia.org/wiki/Microsoft_Edge#Release_history.
+      //
+      // Versions before 15.15063 may contain a JIT bug affecting ES6
+      // constructors (see #161).
+      (browser.name === 'Edge' &&
+       (majorVersion > 15 || (majorVersion === 15 && minorVersion >= 15063))) ||
       (browser.name === 'Firefox' && majorVersion >= 51);
   return !supportsES2015;
 }