+++ CHANGELOG.md
@@ -7,9 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-### Fixed
-* When directory paths don't end in `/`, redirect to the right path, not a
-  filesystem path. https://github.com/Polymer/polyserve/issues/96
+* Add auto-compile support for the Mobile Safari browser.
+
+## [0.18.0](https://github.com/PolymerLabs/polyserve/tree/v0.18.0) (2017-04-18)
+
+* When compiling to ES5, inject the Custom Elements ES5 Adapter into any HTML
+  file where the web components polyfill is included (typically the entry
+  point). This adapter is needed when serving ES5 to browsers that support the
+  native Custom Elements API
+  ([#164](https://github.com/Polymer/polyserve/issues/164)).
+* A server can now be started with a custom entry point; previously index.html
+  was always assumed ([#161](https://github.com/Polymer/polyserve/issues/161)).
+
+## [0.17.0](https://github.com/PolymerLabs/polyserve/tree/v0.17.0) (2017-04-13)
+
+* Add auto-compile support for the Chromium browser.
+* Add auto-compile support for the Opera browser.
+* Require ES5 compilation for Edge < 40 ([#161](https://github.com/Polymer/polyserve/issues/161)).
+* Fixed issue where directory paths redirected to the file system ([#96](https://github.com/Polymer/polyserve/issues/96)).
+* [Breaking] Remove Node v4 support: Node v4 is no longer in Active LTS, so as per the [Polymer Tools Node.js Support Policy](https://www.polymer-project.org/2.0/docs/tools/node-support) polyserve will not support Node v4. Please update to Node v6 or later to continue using the latest verisons of Polymer tooling.
 
 ## [0.16.0](https://github.com/PolymerLabs/polyserve/tree/v0.16.0) (2017-02-14)